add new permissions for admin api and user api

+++ app/Http/Controllers/Api/Admin/AdminApiController.php
@@ -25,12 +25,16 @@ class  AdminApiController extends Controller
      */
     public function index()
     {
-        $admins = $this->admin->list();
+        if(auth()->user()->can('show-admin-information')){
+            $admins = $this->admin->list();
 
-        return $this->success([
-            'data' => AdminResource::collection($admins),
-            'message' => 'Show Admins successfully',
-        ], 200);
+            return $this->success([
+                'data' => AdminResource::collection($admins),
+                'message' => 'Show Admins successfully',
+            ], 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 
     /**
@@ -38,18 +42,22 @@ class  AdminApiController extends Controller
      */
     public function store(StoreAdminRequest $request)
     {
-        $request->validated($request->all());
-        $admin = $this->admin->create([
-            'adm_name' => $request->adm_name,
-            'adm_email' => $request->adm_email,
-            'adm_phone' => $request->adm_phone,
-            'adm_password' => Hash::make($request->adm_password),
-            'adm_role' => $request->adm_role,
-        ]);
-        return $this->success([
-            'data' => new AdminResource($admin),
-            'message' => 'Admin created successfully',
-        ], 200);
+        if(auth()->user()->can('add-new-admin')){
+            $request->validated($request->all());
+            $admin = $this->admin->create([
+                'adm_name' => $request->adm_name,
+                'adm_email' => $request->adm_email,
+                'adm_phone' => $request->adm_phone,
+                'adm_password' => Hash::make($request->adm_password),
+                'adm_role' => $request->adm_role,
+            ]);
+            return $this->success([
+                'data' => new AdminResource($admin),
+                'message' => 'Admin created successfully',
+            ], 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 
     /**
@@ -57,11 +65,15 @@ class  AdminApiController extends Controller
      */
     public function show(string $id)
     {
-        $admin = $this->admin->show($id);
-        return $this->success([
-            'data' => new AdminResource($admin),
-            'message' => 'Show Admin successfully',
-        ], 201);
+        if(auth()->user()->can('show-admin-details-information')){
+            $admin = $this->admin->show($id);
+            return $this->success([
+                'data' => new AdminResource($admin),
+                'message' => 'Show Admin successfully',
+            ], 201);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 
     /**
@@ -69,15 +81,19 @@ class  AdminApiController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $update = $this->admin->update($request->all(), $id);
-        if ($update) {
-            $admin = $this->admin->show($id);
-            return $this->success([
-                'data' => new AdminResource($admin),
-                'message' => 'Admin updated successfully',
-            ], 200);
-        } else {
-            return $this->error('','Admin not updated',400);
+        if(auth()->user()->can('update-admin-information')){
+            $update = $this->admin->update($request->all(), $id);
+            if ($update) {
+                $admin = $this->admin->show($id);
+                return $this->success([
+                    'data' => new AdminResource($admin),
+                    'message' => 'Admin updated successfully',
+                ], 200);
+            } else {
+                return $this->error('','Admin not updated',400);
+            }
+        }else{
+            abort(403, 'You need permission to do this action.');
         }
     }
 
@@ -86,10 +102,14 @@ class  AdminApiController extends Controller
      */
     public function destroy(string $id)
     {
-        $admin = $this->admin->delete($id);
-        return $this->success([
-            'data' => null,
-            'message' => 'Admin deleted successfully',
-        ], 200);
+        if(auth()->user()->can('delete-admin')){
+            $admin = $this->admin->delete($id);
+            return $this->success([
+                'data' => null,
+                'message' => 'Admin deleted successfully',
+            ], 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 }
