+++ app/Http/Controllers/Api/User/UserApiController.php
@@ -26,8 +26,12 @@ class UserApiController extends Controller
      */
     public function index()
     {
-        $users = $this->user->list();
-        return $users;
+        if(auth()->user()->can('show-users-information')){
+            $users = $this->user->list();
+            return $users;
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 
     /**
@@ -35,24 +39,28 @@ class UserApiController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
-        $request->validated($request->all());
-        $user = $this->user->create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'type' => $request->type,
-            'phone' => $request->phone,
-            'title' => $request->title,
-            'company_id' => $request->company_id,
-            'is_first_login' => 1,
-        ]);
-        if ($user) {
-            Mail::to($user->email)->send(new SendCreateMail($user->email,$request->password, $user->name));
+        if(auth()->user()->can('add-new-users')){
+            $request->validated($request->all());
+            $user = $this->user->create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => Hash::make($request->password),
+                'type' => $request->type,
+                'phone' => $request->phone,
+                'title' => $request->title,
+                'company_id' => $request->company_id,
+                'is_first_login' => 1,
+            ]);
+            if ($user) {
+                Mail::to($user->email)->send(new SendCreateMail($user->email,$request->password, $user->name));
+            }
+            return $this->success([
+                'data' => new UserResource($user),
+                'message' => 'User created successfully',
+            ], 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
         }
-        return $this->success([
-            'data' => new UserResource($user),
-            'message' => 'User created successfully',
-        ], 200);
     }
 
     /**
@@ -60,11 +68,15 @@ class UserApiController extends Controller
      */
     public function show(string $id)
     {
-        $user = $this->user->show($id);
-        return $this->success([
-            'data' => new UserResource($user),
-            'message' => null,
-        ], 201);
+        if(auth()->user()->can('show-users-details-information')){
+            $user = $this->user->show($id);
+            return $this->success([
+                'data' => new UserResource($user),
+                'message' => null,
+            ], 201);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
 
     /**
@@ -72,16 +84,20 @@ class UserApiController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $update = $this->user->update($request->all(), $id);
-        if($update){
-            $user = $this->user->show($id);
-            return $this->success([
-                'data' => new UserResource($user),
-                'message' => 'User updated successfully',
-            ], 200);
-        }
-        else{
-            return $this->error(null,'User not updated',400);
+        if(auth()->user()->can('update-user-information')){
+            $update = $this->user->update($request->all(), $id);
+            if($update){
+                $user = $this->user->show($id);
+                return $this->success([
+                    'data' => new UserResource($user),
+                    'message' => 'User updated successfully',
+                ], 200);
+            }
+            else{
+                return $this->error(null,'User not updated',400);
+            }
+        }else{
+            abort(403, 'You need permission to do this action.');
         }
 
     }
@@ -91,11 +107,15 @@ class UserApiController extends Controller
      */
     public function destroy(string $id)
     {
-        $user = $this->user->delete($id);
-        return $this->success([
-            'data' => null,
-            'message' => 'User deleted successfully',
-        ], 200);
+        if(auth()->user()->can('delete-user')){
+            $user = $this->user->delete($id);
+            return $this->success([
+                'data' => null,
+                'message' => 'User deleted successfully',
+            ], 200);
+        }else{
+            abort(403, 'You need permission to do this action.');
+        }
     }
     public function CompanyUsers(string $company_id)
     {
